TodoApp/requirements.txt
aiofiles==25.1.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.1
caio==0.9.24
certifi==2025.11.12
click==8.3.1
colorama==0.4.6
cssbeautifier==1.15.4
djlint==1.36.4
ecdsa==0.19.1
EditorConfig==0.17.1
fastapi==0.121.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jsbeautifier==1.15.4
json5==0.12.1
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pluggy==1.6.0
psycopg2==2.9.11
pyasn1==0.6.1
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.1
pytest-asyncio==1.3.0
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
regex==2025.11.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
python-dotenv==1.2.1
